refactor: service structure
