+++ src/registration/phone_number/mod.rs
@@ -1,0 +1,5 @@
+pub mod router;
+pub mod service;
+
+pub use router::phone_router;
+pub use service::PhoneRegistrationService;
