+++ src/registration/username/mod.rs
@@ -1,0 +1,5 @@
+pub mod router;
+pub mod service;
+
+pub use router::router;
+pub use service::RegistrationService;
