+++ src/registration/username/router.rs
@@ -9,7 +9,7 @@ use utoipa_axum::{router::OpenApiRouter, routes};
 
 use uuid::Uuid;
 
-use super::entities::RegistrationChallenge;
+use crate::registration::entities::RegistrationChallenge;
 use crate::context::AppContext;
 
 const REGISTRATION_TAG: &str = "registration";
